Updated normalization conversions, updated EER diagram, improved record removal auditing checks.

+++ R/qc_remove_record.R
@@ -58,7 +58,9 @@ qc_remove_record <- function(df, tbl_name, reset_extraction = FALSE){
                       qc_flags = paste0('Removed due to foreign key association to removed record in ', tbl_name, 
                                         ' table (', toString(del_ids[[tolower(tbl_name)]]), ")"))
     } else {
-      del_df = df
+      del_df = df %>%
+        dplyr::mutate(qc_notes = paste0("Removed ", tbl_name, " id ", toString(del_ids[[tolower(tbl_name)]])),
+                      qc_flags = paste0("Removed ", tbl_name, " id ", toString(del_ids[[tolower(tbl_name)]])))
     }
     
     # Skip deleting document records if just resetting extraction
@@ -116,7 +118,19 @@ qc_remove_record <- function(df, tbl_name, reset_extraction = FALSE){
     db_update_tbl(df = del_df, tblName = del_tbl)
     # twice to audit QC'd record
     db_update_tbl(df = del_df, tblName = del_tbl)
-    # Delete database entries
-    db_query_cvt(paste0("DELETE FROM cvt.", del_tbl, " WHERE id IN (", toString(del_df$id), ")"))
+    # Check if audit worked
+    curr_rec = db_query_cvt(paste0("SELECT id, version FROM cvt.", tbl_name, 
+                                   " WHERE id = ", del_df$id))
+    audit_check = db_query_cvt(paste0("SELECT * FROM cvt.cvt_audit WHERE fk_table_id = ", 
+                                      del_df$id, " AND fk_table_name = '", tbl_name,"' AND ",
+                                      "version = ", curr_rec$version - 1))
+    # If audit successful, delete the record
+    if(nrow(audit_check)){
+      # Delete database entries
+      db_query_cvt(paste0("DELETE FROM cvt.", del_tbl, " WHERE id IN (", toString(del_df$id), ")"))  
+    } else {
+      stop("Record not audited as expected...record not deleted...")
+    }
+    
   }
 }
